requirements.txt
langchain_google_genai
streamlit
python-dotenv
pdf2image
faiss-cpu
google-generativeai
langchain
PyPDF2
chromadb
